Scoping offer objects & generating context
Implements checks to database api to scope calls to ensure proper access
to objects. Uses oslo_context to pass a project_id to object creation,
and when querying database, checks that oslo_context project_id matches
with the object's project_id or that the context belongs to an admin.
Also modifies apis and business objects for bids and contracts to pass
in a context.
TG-163